requirements.txt
aiofiles==23.1.0
aiohttp==3.8.4
aiosignal==1.3.1
altair==4.2.2
anyio==3.6.2
appnope==0.1.3
asttokens==2.1.0
async-generator==1.10
async-timeout==4.0.2
attrs==22.2.0
backcall==0.2.0
beautifulsoup4==4.12.0
certifi==2022.9.14
chardet==3.0.4
charset-normalizer==2.1.1
click==8.1.3
contourpy==1.0.5
cycler==0.11.0
debugpy==1.6.3
decorator==5.1.1
discord.py==2.2.2
dpkt==1.9.8
entrypoints==0.4
exceptiongroup==1.1.1
executing==1.2.0
fastapi==0.95.1
ffmpy==0.3.0
filelock==3.12.0
Flask==2.2.3
fonttools==4.37.4
frozenlist==1.3.3
fsspec==2023.4.0
gradio==3.27.0
gradio_client==0.1.3
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2022.9.1
httpcore==0.9.1
httpx==0.13.3
huggingface-hub==0.13.4
hyperframe==5.2.0
idna==2.10
ipykernel==6.17.1
ipython==8.6.0
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.17.3
jupyter_client==7.4.5
jupyter_core==5.0.0
kiwisolver==1.4.4
linkify-it-py==2.0.0
lxml==4.9.2
markdown-it-py==2.2.0
MarkupSafe==2.1.2
matplotlib==3.6.1
matplotlib-inline==0.1.6
mdit-py-plugins==0.3.3
mdurl==0.1.2
multidict==6.0.4
nest-asyncio==1.5.6
networkx==3.1
numpy==1.23.4
openai==0.27.4
OpenCC==0.2
opencc-python-reimplemented==0.1.7
opencv-contrib-python==4.6.0.66
orjson==3.8.10
outcome==1.2.0
packaging==21.3
pandas==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
platformdirs==2.5.4
praw==7.6.0
prawcore==2.3.0
prompt-toolkit==3.0.32
protobuf==4.21.9
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pybase64==1.2.3
pydantic==1.10.7
pydub==0.25.1
Pygments==2.13.0
pyparsing==3.0.9
pyrsistent==0.19.3
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2022.4
PyYAML==6.0
pyzmq==24.0.1
requests==2.28.1
rfc3986==1.5.0
scikit-learn==1.1.2
scipy==1.9.2
selenium==4.8.2
semantic-version==2.10.0
six==1.16.0
sklearn==0.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4
stack-data==0.6.1
starlette==0.26.1
threadpoolctl==3.1.0
toolz==0.12.0
torch==1.13.0
torchvision==0.14.0
tornado==6.2
tqdm==4.64.1
traitlets==5.5.0
trio==0.22.0
trio-websocket==0.10.2
typing_extensions==4.4.0
uc-micro-py==1.0.1
update-checker==0.18.0
urllib3==1.26.12
uvicorn==0.21.1
wcwidth==0.2.5
websocket-client==1.4.1
websockets==11.0.2
Werkzeug==2.2.3
wsproto==1.2.0
yarl==1.8.2
